Create Object interaction
In the exercise both the main program and the class "Calculator" have been completed. Your task is to make up class "Printer" that asks the user for two integers, calls for the class "Calculator" method "Sum()" transferring the integers received as input to the method. Method "Sum()" calculates the sum of the integers and returns the result. At the end method "Print()" prints the given sum on screen. I.e. the code below doesnt need to be copied to the answer field.